changements stats en long au lieu d'integer

+++ src/main/java/qengine/storage/ThrdValue.java
@@ -3,9 +3,9 @@ package qengine.storage;
 import java.util.Set;
 
 public class ThrdValue {
-    Integer stat;
+    long stat;
     Set<Integer> set;
-    public ThrdValue(Integer stat) {
+    public ThrdValue(long stat) {
         this.stat = stat;
     }
     public ThrdValue(Set<Integer> set) {
